+++ README.md
@@ -4,10 +4,11 @@
 
 
 #### - Description -
-・ユーザー登録・ログイン、文章投稿、他のユーザーの投稿にコメントすることができるSNSサイト<br />
-・フォローしたユーザーの投稿のみをホーム画面に表示
+・ユーザー登録・ログイン、文章投稿、他のユーザーの投稿にコメント、投稿へのいいねをすることができるSNSサイト<br />
+・フォローしたユーザー、自分の投稿のみをホーム画面に表示
 
 ・React<br />
+・TypeScript<br />
 ・TailwindCSS<br />
 ・PostgreSQL<br />
 ・Node.js<br />
