+++ package.json
@@ -18,6 +18,7 @@
     "@types/react-modal": "^3.13.1",
     "axios": "^1.2.2",
     "dayjs": "^1.11.7",
+    "framer-motion": "^9.0.1",
     "react": "^18.2.0",
     "react-custom-scrollbars": "^4.2.1",
     "react-dom": "^18.2.0",
