--- src/DummyData.js
@@ -1,213 +1,0 @@
-export const Users = [
-    {
-      id: 1,
-      profilePicture: "/person/1.jpeg",
-      username: "mai",
-    },
-    {
-      id: 2,
-      profilePicture: "/person/2.jpeg",
-      username: "Yamaki",
-    },
-    {
-      id: 3,
-      profilePicture: "/person/3.jpeg",
-      username: "Koga",
-    },
-    {
-      id: 4,
-      profilePicture: "/person/4.jpeg",
-      username: "Matukubo",
-    },
-    {
-      id: 5,
-      profilePicture: "/person/5.jpeg",
-      username: "Kikukawa",
-    },
-  ];
-
-  export const AllPosts = [
-    {
-      id: 1,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 2,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 3,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 4,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 5,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 6,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 7,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-  ]
-  
-  export const FollowPosts = [
-    {
-      id: 1,
-      desc: "おはようございます",
-      photo: "/post/1.jpeg",
-      date: "5分前",
-      username: "田中",
-      userId: "user_sns",
-      like: 20,
-      comment: 4,
-    },
-    {
-      id: 2,
-      desc: "SNS作成中です",
-      photo: "/post/2.jpeg",
-      date: "2分前",
-      username: "山田",
-      userId: "user_sns",
-      like: 140,
-      comment: 13,
-    },
-    {
-      id: 3,
-      desc: "今日はいい天気です",
-      photo: "/post/3.jpeg",
-      date: "10分前",
-      username: "佐藤",
-      userId: "user_sns",
-      like: 230,
-      comment: 29,
-    },
-    {
-      id: 4,
-      photo: "/post/4.jpeg",
-      date: "1分前",
-      username: "ピカチュウ",
-      userId: "user_sns",
-      like: 10,
-      comment: 1,
-    },
-    {
-      id: 5,
-      desc: "おにぎりを作りました",
-      photo: "/post/5.jpeg",
-      date: "1分前",
-      username: "おにぎり",
-      userId: "user_sns",
-      like: 100,
-      comment: 3,
-    },
-  ];
-  
-  export const MyPosts = [
-    {
-      id: 1,
-      desc: "2023年になりました",
-      date: "10分前",
-      username: "mai",
-      userId: "mai_sns",
-      like: 1,
-      comment: 1,
-    },
-    {
-      id: 2,
-      desc: "2023年になりました",
-      date: "10分前",
-      username: "mai",
-      userId: "mai_sns",
-      like: 1,
-      comment: 1,
-    },
-    {
-      id: 3,
-      desc: "2023年になりました",
-      date: "10分前",
-      username: "mai",
-      userId: "mai_sns",
-      like: 1,
-      comment: 1,
-    },
-  ]
-
-
-  export const LikePosts = [
-    {
-      id: 1,
-      desc: "2023年になりました",
-      date: "10分前",
-      username: "mai",
-      userId: "mai_sns",
-      like: 1,
-      comment: 1,
-    },
-    {
-      id: 2,
-      desc: "2023年になりました",
-      date: "10分前",
-      username: "mai",
-      userId: "mai_sns",
-      like: 1,
-      comment: 1,
-    },
-    {
-      id: 3,
-      desc: "2023年になりました",
-      date: "10分前",
-      username: "mai",
-      userId: "mai_sns",
-      like: 1,
-      comment: 1,
-    },
-  ]
