+++ src/components/Timeline.tsx
@@ -7,7 +7,7 @@ import { Posts, UserData } from "../types/Types";
 
 const Timeline: React.FC = () => {
   const [posts, setPosts] = useState<Posts[]>([]);
-  const [userData, setUserData] = useState<any>([]); 
+  const [userData, setUserData] = useState<any>([]);
   const data: string | null = localStorage.getItem("id");
 
   useEffect(() => {
@@ -31,26 +31,28 @@ const Timeline: React.FC = () => {
   }, [posts]);
 
   const followUser: string[] = userData.followings?.map(
-    (user: any) => user.userId
+    (user: UserData) => user.userId
   );
+
   followUser?.push(userData.userId);
 
   const newPost: Posts[] = posts.filter((item: Posts) =>
     followUser?.includes(item.userId)
   );
 
-  console.log(newPost,9)
   return (
-      <div className="bg-white basis-2/4 max-h-screen overflow-scroll">
-        <div>
-          <Share userData={userData} />
-          {newPost.map((followpost: Posts) => {
-            return (
+    <div className="bg-white basis-2/4 max-h-screen overflow-scroll">
+      <div>
+        <Share userData={userData} />
+        {newPost.map((followpost: Posts) => {
+          return (
+            <React.Fragment key={followpost.id}>
               <Post post={followpost} key={followpost.id} userData={userData} />
-            );
-          })}
-        </div>
+            </React.Fragment>
+          );
+        })}
       </div>
+    </div>
   );
 };
