+++ src/components/profile/Follow.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react";
 import { Link, useLocation, useNavigate, useParams } from "react-router-dom";
 import { FollowItem } from "../../types/Types";
 
-const Follow = () => {
+const Follow: React.FC = () => {
   const navigate = useNavigate();
   const params = useParams();
 
@@ -18,13 +18,11 @@ const Follow = () => {
     navigate(`/profile/${params.id}`);
   };
 
-  console.log(followList)
-
   return (
-    <div>
+    <div className="max-h-screen overflow-scroll">
       <div>
         <button onClick={clickReturn}>
-          <ArrowBack />
+          <ArrowBack className="m-5"/>
         </button>
       </div>
       <div className="bg-mygreen w-4/12 my-6 mx-auto p-3 rounded-3xl">
@@ -33,7 +31,11 @@ const Follow = () => {
           {followList?.map((followUser) => {
             return (
               <React.Fragment key={followUser.userId}>
-                <li className="ml-2 mb-2"><Link to={`/profile/${followUser.followerId}`}>{followUser.userId}</Link></li>
+                <li className="ml-2 mb-2">
+                  <Link to={`/profile/${followUser.followerId}`}>
+                    {followUser.userId}
+                  </Link>
+                </li>
               </React.Fragment>
             );
           })}
