+++ src/components/rightbar/Rightbar.tsx
@@ -5,7 +5,7 @@ const Rightbar: React.FC = () => {
   return (
     <div className="bg-white basis-1/4">
       <div className="bg-mygreen w-10/12 my-0 mx-auto max-h-screen rounded-3xl">
-      <SearchInput />
+        <SearchInput />
       </div>
     </div>
   );
